chore: support old legacy urls
