+++ src/AppWrapper.js
@@ -1,8 +1,9 @@
 import { CachedDataQueryProvider } from '@dhis2/analytics'
 import { D2Shim } from '@dhis2/app-runtime-adapter-d2'
 import { DataStoreProvider } from '@dhis2/app-service-datastore'
-import { CenteredContent, CircularLoader } from '@dhis2/ui'
+import { CssVariables, CenteredContent, CircularLoader } from '@dhis2/ui'
 import log from 'loglevel'
+import queryString from 'query-string'
 import React from 'react'
 import { Provider as ReduxProvider } from 'react-redux'
 import App from './components/app/App.js'
@@ -12,6 +13,7 @@ import store from './store/index.js'
 import { USER_DATASTORE_NAMESPACE } from './util/analyticalObject.js'
 import { appQueries, providerDataTransformation } from './util/app.js'
 import './locales/index.js'
+import history from './util/history.js'
 
 log.setLevel(
     process.env.NODE_ENV === 'production' ? log.levels.INFO : log.levels.TRACE
@@ -39,6 +41,39 @@ const d2Config = {
 }
 
 const AppWrapper = () => {
+    // support legacy urls
+    const queryParams = queryString.parse(window.location.search, {
+        parseBooleans: true,
+    })
+    const [base] = window.location.href.split('?')
+
+    if (queryParams.id) {
+        // /?id=ytkZY3ChM6J
+        // /?id=ZBjCfSaLSqD&interpretationid=yKqhXZdeJ6a
+        let newPath = queryParams.id
+
+        let interpretationId
+        if (queryParams.interpretationId) {
+            interpretationId = queryParams.interpretationId
+        } else if (queryParams.interpretationid) {
+            interpretationId = queryParams.interpretationid
+        }
+
+        if (interpretationId) {
+            newPath += `?interpretationId=${interpretationId}`
+        }
+
+        // replace history && hash history
+        window.history.replaceState({}, '', `${base}#/${newPath}`)
+        history.replace(`/${newPath}`)
+    } else if (queryParams.currentAnalyticalObject === true) {
+        // /?currentAnalyticalObject=true
+
+        // replace history && hash history
+        window.history.replaceState({}, '', `${base}#/currentAnalyticalObject`)
+        history.replace('/currentAnalyticalObject')
+    }
+
     return (
         <ReduxProvider store={store}>
             <DataStoreProvider namespace={USER_DATASTORE_NAMESPACE}>
@@ -67,6 +102,7 @@ const AppWrapper = () => {
                             >
                                 <WindowDimensionsProvider>
                                     <OrgUnitsProvider>
+                                        <CssVariables colors spacers theme />
                                         <App />
                                     </OrgUnitsProvider>
                                 </WindowDimensionsProvider>
