+++ src/actions/analyticalObject.js
@@ -1,11 +1,4 @@
-import log from 'loglevel'
 import * as types from '../constants/actionTypes.js'
-import {
-    clearAnalyticalObjectFromUrl,
-    hasSingleDataDimension,
-    getThematicLayerFromAnalyticalObject,
-} from '../util/analyticalObject.js'
-import { addLayer } from './layers.js'
 
 export const setAnalyticalObject = (ao) => ({
     type: types.ANALYTICAL_OBJECT_SET,
@@ -15,17 +8,3 @@ export const setAnalyticalObject = (ao) => ({
 export const clearAnalyticalObject = () => ({
     type: types.ANALYTICAL_OBJECT_CLEAR,
 })
-
-export const tSetAnalyticalObject = (ao) => async (dispatch) => {
-    try {
-        clearAnalyticalObjectFromUrl()
-        return hasSingleDataDimension(ao)
-            ? getThematicLayerFromAnalyticalObject(ao).then((layer) =>
-                  dispatch(addLayer(layer))
-              )
-            : dispatch(setAnalyticalObject(ao))
-    } catch (e) {
-        log.error('Could not load current analytical object')
-        return e
-    }
-}
