+++ src/components/app/App.js
@@ -1,155 +1,82 @@
-import { useCachedDataQuery } from '@dhis2/analytics'
-import { useDataEngine } from '@dhis2/app-runtime'
-import { useAlert } from '@dhis2/app-service-alerts'
-import { useSetting } from '@dhis2/app-service-datastore'
-import i18n from '@dhis2/d2-i18n'
-import { CssVariables } from '@dhis2/ui'
-import queryString from 'query-string'
-import React, { useEffect, useCallback } from 'react'
-import { useDispatch } from 'react-redux'
-import { tSetAnalyticalObject } from '../../actions/analyticalObject.js'
-import { setInterpretation } from '../../actions/interpretations.js'
-import { tOpenMap, newMap } from '../../actions/map.js'
-import {
-    ALERT_CRITICAL,
-    ALERT_MESSAGE_DYNAMIC,
-    // ALERT_OPTIONS_DYNAMIC,
-    // ALERT_SUCCESS_DELAY,
-} from '../../constants/alerts.js'
-import { CURRENT_AO_KEY } from '../../util/analyticalObject.js'
-import history from '../../util/history.js'
-import AppLayout from './AppLayout.js'
+import cx from 'classnames'
+import React, { useState } from 'react'
+import { useSelector } from 'react-redux'
+import AlertStack from '../alerts/AlertStack.js'
+import BottomPanel from '../datatable/BottomPanel.js'
+import DownloadModeMenu from '../download/DownloadMenubar.js'
+import DownloadSettings from '../download/DownloadSettings.js'
+import LayerEdit from '../edit/LayerEdit.js'
+import LayersPanel from '../layers/LayersPanel.js'
+import LayersLoader from '../loaders/LayersLoader.js'
+import ContextMenu from '../map/ContextMenu.js'
+import MapPosition from '../map/MapPosition.js'
+import OpenAsMapDialog from '../openAs/OpenAsMapDialog.js'
+import AppMenu from './AppMenu.js'
+import DetailsPanel from './DetailsPanel.js'
 import './App.css'
-import './styles/App.module.css'
+import styles from './styles/App.module.css'
+import { useLoadMap } from './useLoadMap.js'
 
-const getUrlParameter = (location, name) => {
-    const parsed = queryString.parse(location.search, { parseBooleans: true })
-    return parsed[name]
-}
+// const getUrlParameter = (location, name) => {
+//     const parsed = queryString.parse(location.search, { parseBooleans: true })
+//     return parsed[name]
+// }
 
-const getMapId = (hashLocation) => {
-    const pathParts = hashLocation.pathname.slice(1).split('/')
-    if (pathParts[0]) {
-        return pathParts[0]
-    }
+// const getMapId = (hashLocation) => {
+//     const pathParts = hashLocation.pathname.slice(1).split('/')
+//     if (pathParts[0]) {
+//         return pathParts[0]
+//     }
 
-    // support /?id=ytkZY3ChM6J for backwards compatibility
-    return getUrlParameter(hashLocation, 'id')
-}
+//     // support /?id=ytkZY3ChM6J for backwards compatibility
+//     return getUrlParameter(hashLocation, 'id')
+// }
 
 const App = () => {
-    // const [previousLocation, setPreviousLocation] = useState()
-    const { systemSettings, basemaps } = useCachedDataQuery()
-    const defaultBasemap = systemSettings.keyDefaultBaseMap
-    const engine = useDataEngine()
-    const [currentAO] = useSetting(CURRENT_AO_KEY)
-    const dispatch = useDispatch()
-    const openMapErrorAlert = useAlert(ALERT_MESSAGE_DYNAMIC, ALERT_CRITICAL)
+    useLoadMap()
+    const [interpretationsRenderCount, setInterpretationsRenderCount] =
+        useState(1)
 
-    const setInterpretationId = useCallback(
-        (location) => {
-            // support both lower and camel case for backwards compatibility
-            const interpretationId =
-                getUrlParameter(location, 'interpretationid') ||
-                getUrlParameter(location, 'interpretationId')
-
-            console.log('jj interpretationId', interpretationId)
-            if (interpretationId) {
-                dispatch(setInterpretation(interpretationId))
-            }
-        },
-        [dispatch]
+    const dataTableOpen = useSelector((state) => !!state.dataTable)
+    const downloadModeOpen = useSelector(
+        (state) => !!state.download.downloadMode
+    )
+    const detailsPanelOpen = useSelector(
+        (state) => state.ui.rightPanelOpen && !state.orgUnitProfile
     )
 
-    const loadMap = useCallback(
-        async (location, isNew) => {
-            console.log('jj location', location)
-            const isExisting = location.pathname.length > 1
-
-            if (isExisting) {
-                const mapId = getMapId(location)
-                if (mapId) {
-                    const error = await dispatch(
-                        tOpenMap({
-                            mapId,
-                            defaultBasemap,
-                            engine,
-                            basemaps,
-                        })
-                    )
-
-                    if (error) {
-                        openMapErrorAlert.show({
-                            msg: i18n.t(
-                                `Error while opening map: ${error.message}`,
-                                {
-                                    nsSeparator: ';',
-                                }
-                            ),
-                        })
-
-                        return
-                    }
-                } else if (
-                    getUrlParameter(location, 'currentAnalyticalObject') ===
-                    'true'
-                ) {
-                    await dispatch(tSetAnalyticalObject(currentAO))
-                }
-            } else if (isNew) {
-                dispatch(newMap())
-            }
-
-            setInterpretationId(location)
-            // setPreviousLocation(location.pathname)
-        },
-        [
-            basemaps,
-            currentAO,
-            defaultBasemap,
-            dispatch,
-            engine,
-            openMapErrorAlert,
-            setInterpretationId,
-        ]
-    )
-
-    useEffect(() => {
-        console.log('jj App useEffect')
-        loadMap(history.location)
-
-        const unlisten = history.listen((myhistory) => {
-            console.log('jj myhistory', myhistory)
-            const { location } = myhistory
-            // const isSaving = location.state?.isSaving
-            // const isOpening = location.state?.isOpening
-            // const isResetting = location.state?.isResetting
-            const isModalOpening = location.state?.isModalOpening
-            const isModalClosing = location.state?.isModalClosing
-            // const isValidLocationChange =
-            // previousLocation !== location.pathname &&
-            // !isModalOpening && !isModalClosing
-
-            // TODO navigation confirm dialog
-
-            if (isModalOpening) {
-                setInterpretationId(location)
-            } else if (isModalClosing) {
-                dispatch(setInterpretation())
-                // } else if (isSaving || isOpening || isResetting) {
-            } else {
-                console.log('jj reload')
-                loadMap(location, location.pathname === '/')
-            }
-        })
-
-        return () => unlisten && unlisten()
-    }, [loadMap, dispatch, setInterpretationId])
+    const onFileMenuAction = () =>
+        detailsPanelOpen &&
+        setInterpretationsRenderCount(interpretationsRenderCount + 1)
 
     return (
         <>
-            <CssVariables colors spacers theme />
-            <AppLayout />
+            {downloadModeOpen ? (
+                <DownloadModeMenu />
+            ) : (
+                <AppMenu onFileMenuAction={onFileMenuAction} />
+            )}
+            <div
+                className={cx(styles.content, {
+                    [styles.downloadContent]: downloadModeOpen,
+                })}
+            >
+                {downloadModeOpen ? <DownloadSettings /> : <LayersPanel />}
+                <div className={styles.appMapAndTable}>
+                    <MapPosition />
+                    {dataTableOpen && <BottomPanel />}
+                </div>
+                {!downloadModeOpen && (
+                    <DetailsPanel
+                        interpretationsRenderCount={interpretationsRenderCount}
+                    />
+                )}
+            </div>
+            <LayersLoader />
+            <ContextMenu />
+            <LayerEdit />
+            <AlertStack />
+            <OpenAsMapDialog />
         </>
     )
 }
