--- src/components/app/AppLayout.js
@@ -1,66 +1,0 @@
-import cx from 'classnames'
-import React, { useState } from 'react'
-import { useSelector } from 'react-redux'
-import AlertStack from '../alerts/AlertStack.js'
-import BottomPanel from '../datatable/BottomPanel.js'
-import DownloadModeMenu from '../download/DownloadMenubar.js'
-import DownloadSettings from '../download/DownloadSettings.js'
-import LayerEdit from '../edit/LayerEdit.js'
-import LayersPanel from '../layers/LayersPanel.js'
-import LayersLoader from '../loaders/LayersLoader.js'
-import ContextMenu from '../map/ContextMenu.js'
-import MapPosition from '../map/MapPosition.js'
-import OpenAsMapDialog from '../openAs/OpenAsMapDialog.js'
-import AppMenu from './AppMenu.js'
-import DetailsPanel from './DetailsPanel.js'
-import styles from './styles/AppLayout.module.css'
-
-const AppLayout = () => {
-    const [interpretationsRenderCount, setInterpretationsRenderCount] =
-        useState(1)
-
-    const dataTableOpen = useSelector((state) => !!state.dataTable)
-    const downloadModeOpen = useSelector(
-        (state) => !!state.download.downloadMode
-    )
-    const detailsPanelOpen = useSelector(
-        (state) => state.ui.rightPanelOpen && !state.orgUnitProfile
-    )
-
-    const onFileMenuAction = () =>
-        detailsPanelOpen &&
-        setInterpretationsRenderCount(interpretationsRenderCount + 1)
-
-    return (
-        <>
-            {downloadModeOpen ? (
-                <DownloadModeMenu />
-            ) : (
-                <AppMenu onFileMenuAction={onFileMenuAction} />
-            )}
-            <div
-                className={cx(styles.content, {
-                    [styles.downloadContent]: downloadModeOpen,
-                })}
-            >
-                {downloadModeOpen ? <DownloadSettings /> : <LayersPanel />}
-                <div className={styles.appMapAndTable}>
-                    <MapPosition />
-                    {dataTableOpen && <BottomPanel />}
-                </div>
-                {!downloadModeOpen && (
-                    <DetailsPanel
-                        interpretationsRenderCount={interpretationsRenderCount}
-                    />
-                )}
-            </div>
-            <LayersLoader />
-            <ContextMenu />
-            <LayerEdit />
-            <AlertStack />
-            <OpenAsMapDialog />
-        </>
-    )
-}
-
-export default AppLayout
