+++ src/components/app/styles/App.module.css
@@ -30,3 +30,19 @@
     background: #cbcdcf;
     border-radius: 3px;
 }
+
+.content {
+    display: flex;
+    flex-direction: row;
+    height: calc(100vh - var(--header-height) - var(--toolbar-height));
+}
+
+.downloadContent {
+    margin-top: var(--header-height);
+    height: calc(100vh - var(--header-height));
+}
+
+.appMapAndTable {
+    flex: auto;
+    position: relative;
+}
