--- src/components/app/styles/AppLayout.module.css
@@ -1,15 +1,0 @@
-.content {
-    display: flex;
-    flex-direction: row;
-    height: calc(100vh - var(--header-height) - var(--toolbar-height));
-}
-
-.downloadContent {
-    margin-top: var(--header-height);
-    height: calc(100vh - var(--header-height));
-}
-
-.appMapAndTable {
-    flex: auto;
-    position: relative;
-}
